// play/chess-tourney-frontend/src/api/model/gameFilterParameter.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * (title)
 * API for a custom three-player chess variant game server. Authentication is required via a JWT Bearer token for protected endpoints.
 * OpenAPI spec version: 0.0.0
 */

export type GameFilterParameter =
  (typeof GameFilterParameter)[keyof typeof GameFilterParameter];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GameFilterParameter = {
  Waiting: "Waiting",
  InProgress: "InProgress",
  Finished: "Finished",
} as const;