// play/chess-tourney-frontend/src/api/model/player.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * (title)
 * API for a custom three-player chess variant game server. Authentication is required via a JWT Bearer token for protected endpoints.
 * OpenAPI spec version: 0.0.0
 */
import type { PlayerRelation } from "./playerRelation";

export interface Player {
  id: number;
  username: string;
  relation: PlayerRelation;
  isOwner: boolean;
  isConnectedToGame: boolean;
}